ref(metrics): Cleans up metrics API
 Hacks out the `MockingDataSource`, and
 moves the `SnubaDataSource` into its own
 module
