+++ src/sentry/snuba/metrics/__init__.py
@@ -1,0 +1,2 @@
+from .datasource import *  # NOQA
+from .helpers import *  # NOQA
