+++ tests/sentry/api/endpoints/test_organization_metrics.py
@@ -1,14 +1,10 @@
 import time
-from copy import deepcopy
 from typing import Optional
-from unittest import mock
 
 from django.urls import reverse
 
 from sentry.models import ApiToken
 from sentry.sentry_metrics import indexer
-from sentry.sentry_metrics.sessions import SessionMetricKey
-from sentry.snuba.metrics import _METRICS
 from sentry.testutils import APITestCase
 from sentry.testutils.cases import SessionMetricsTestCase
 from sentry.testutils.helpers import with_feature
@@ -54,222 +50,29 @@ class OrganizationMetricsPermissionTest(APITestCase):
             assert response.status_code == 404
 
 
-class OrganizationMetricsTest(APITestCase):
-
-    endpoint = "sentry-api-0-organization-metrics-index"
-
-    def setUp(self):
-        super().setUp()
-        self.login_as(user=self.user)
-
-    @with_feature(FEATURE_FLAG)
-    def test_response(self):
-        response = self.get_valid_response(self.project.organization.slug)
-
-        required_fields = {"name", "operations", "type"}
-        optional_fields = {"unit"}
-
-        for item in response.data:
-
-            # All required fields are there:
-            assert required_fields <= item.keys()
-
-            # Only optional field is unit:
-            additional_fields = item.keys() - required_fields
-            if additional_fields:
-                assert additional_fields <= optional_fields
-
-
-class OrganizationMetricDetailsTest(APITestCase):
-
-    endpoint = "sentry-api-0-organization-metric-details"
-
-    def setUp(self):
-        super().setUp()
-        self.login_as(user=self.user)
-
-    @with_feature(FEATURE_FLAG)
-    def test_unknown_metric(self):
-        response = self.get_response(self.project.organization.slug, "foo")
-
-        assert response.status_code == 404
-
-    @with_feature(FEATURE_FLAG)
-    def test_valid_response(self):
-
-        response = self.get_success_response(
-            self.project.organization.slug, SessionMetricKey.SESSION.value
-        )
-
-        assert response.data["name"] == SessionMetricKey.SESSION.value
-        assert "tags" in response.data
-        assert all(isinstance(item, str) for item in response.data["tags"])
-
-
-_EXTENDED_METRICS = deepcopy(_METRICS)
-_EXTENDED_METRICS[SessionMetricKey.USER.value]["tags"] = dict(
-    _EXTENDED_METRICS[SessionMetricKey.USER.value]["tags"], custom_user_tag=[""]
-)
-_EXTENDED_METRICS[SessionMetricKey.SESSION.value]["tags"] = dict(
-    _EXTENDED_METRICS[SessionMetricKey.SESSION.value]["tags"], custom_session_tag=["foo", "bar"]
-)
-
-
-class OrganizationMetricTagsTest(APITestCase):
-
-    endpoint = "sentry-api-0-organization-metrics-tags"
-
-    def setUp(self):
-        super().setUp()
-        self.login_as(user=self.user)
-
-    @with_feature(FEATURE_FLAG)
-    @mock.patch("sentry.snuba.metrics._METRICS", _EXTENDED_METRICS)
-    def test_response(self):
-
-        response = self.get_success_response(self.project.organization.slug)
-
-        # Check if data are sane:
-        assert isinstance(response.data, list)
-        assert all(isinstance(item, dict) for item in response.data)
-
-        # Check if intersection works:
-        tags = {tag["key"] for tag in response.data}
-        assert "environment" in tags
-        assert "custom_session_tag" in tags  # from 'session' tags
-        assert "custom_user_tag" in tags  # from 'user' tags
-
-    @with_feature(FEATURE_FLAG)
-    @mock.patch("sentry.snuba.metrics._METRICS", _EXTENDED_METRICS)
-    def test_filtered_response(self):
-
-        response = self.get_success_response(
-            self.project.organization.slug, metric=SessionMetricKey.SESSION.value
-        )
-
-        # Check that only tags from this metrics appear:
-        tags = {tag["key"] for tag in response.data}
-        assert "environment" in tags
-        assert "custom_session_tag" in tags  # from 'session' tags
-        assert "custom_user_tag" not in tags  # from 'user' tags
-
-    @with_feature(FEATURE_FLAG)
-    def test_two_filters(self):
-
-        response = self.get_success_response(
-            self.project.organization.slug,
-            metric=[SessionMetricKey.USER.value, SessionMetricKey.SESSION.value],
-        )
-
-        # Check that only tags from this metrics appear:
-        tags = {tag["key"] for tag in response.data}
-        assert "environment" in tags
-        assert "custom_session_tag" not in tags  # from 'session' tags
-        assert "custom_user_tag" not in tags  # from 'user' tags
-
-    @with_feature(FEATURE_FLAG)
-    def test_bad_filter(self):
-        response = self.get_response(self.project.organization.slug, metric="bad")
-
-        assert response.status_code == 400
-
-
-class OrganizationMetricTagDetailsTest(APITestCase):
-
-    endpoint = "sentry-api-0-organization-metrics-tag-details"
-
-    def setUp(self):
-        super().setUp()
-        self.login_as(user=self.user)
-
-    @with_feature(FEATURE_FLAG)
-    def test_unknown_tag(self):
-        response = self.get_success_response(self.project.organization.slug, "bar")
-
-        assert response.data == []
-
-    @with_feature(FEATURE_FLAG)
-    def test_existing_tag(self):
-        response = self.get_valid_response(self.project.organization.slug, "environment")
-
-        assert response.status_code == 200
-
-        # Check if data are sane:
-        assert isinstance(response.data, list)
-        for item in response.data:
-            assert isinstance(item, dict), item
-
-        assert "production" in {tag["value"] for tag in response.data}
-
-    @with_feature(FEATURE_FLAG)
-    @mock.patch("sentry.snuba.metrics._METRICS", _EXTENDED_METRICS)
-    def test_filtered_response(self):
-
-        response = self.get_success_response(
-            self.project.organization.slug,
-            "custom_session_tag",
-            metric=SessionMetricKey.SESSION.value,
-        )
-
-        # Check that only tags from this metrics appear:
-        assert {tag["value"] for tag in response.data} == {"foo", "bar"}
-
-    @with_feature(FEATURE_FLAG)
-    def test_two_filters(self):
-
-        response = self.get_success_response(
-            self.project.organization.slug,
-            "environment",
-            metric=[SessionMetricKey.USER.value, SessionMetricKey.SESSION.value],
-        )
-
-        assert {tag["value"] for tag in response.data} == {"production", "staging"}
-
-    @with_feature(FEATURE_FLAG)
-    def test_bad_filter(self):
-        response = self.get_response(self.project.organization.slug, "environment", metric="bad")
-
-        assert response.status_code == 400
-
-
-class OrganizationMetricDataTest(APITestCase):
-
+class OrganizationMetricDataTest(SessionMetricsTestCase, APITestCase):
     endpoint = "sentry-api-0-organization-metrics-data"
 
     def setUp(self):
         super().setUp()
+        self.project2 = self.create_project()
         self.login_as(user=self.user)
 
     @with_feature(FEATURE_FLAG)
     def test_missing_field(self):
-        response = self.get_response(
-            self.project.organization.slug,
-        )
-
+        response = self.get_response(self.project.organization.slug)
         assert response.status_code == 400
+        assert response.json()["detail"] == 'Request is missing a "field"'
 
     @with_feature(FEATURE_FLAG)
     def test_invalid_field(self):
-
         for field in ["", "(*&%", "foo(session", "foo(session)", "sum(bar)"]:
             response = self.get_response(self.project.organization.slug, field=field)
-
             assert response.status_code == 400
 
     @with_feature(FEATURE_FLAG)
-    def test_valid_operation(self):
-        response = self.get_response(
-            self.project.organization.slug, field="sum(sentry.sessions.session)"
-        )
-
-        assert response.status_code == 200
-
-        # Only one group:
-        groups = response.data["groups"]
-        assert len(groups) == 1 and groups[0]["by"] == {}
-
-    @with_feature(FEATURE_FLAG)
     def test_groupby_single(self):
+        indexer.record("environment")
         response = self.get_response(
             self.project.organization.slug,
             field="sum(sentry.sessions.session)",
@@ -279,37 +82,20 @@ class OrganizationMetricDataTest(APITestCase):
         assert response.status_code == 200
 
     @with_feature(FEATURE_FLAG)
-    def test_groupby_multiple(self):
+    def test_invalid_filter(self):
+        query = "release:foo or "
         response = self.get_response(
             self.project.organization.slug,
             field="sum(sentry.sessions.session)",
-            groupBy=["environment", "session.status"],
+            groupBy="environment",
+            query=query,
         )
-
-        assert response.status_code == 200
-
-        groups = response.data["groups"]
-        assert len(groups) >= 2 and all(
-            group["by"].keys() == {"environment", "session.status"} for group in groups
-        )
-
-    @with_feature(FEATURE_FLAG)
-    def test_invalid_filter(self):
-        for query in [
-            "release:foo or ",
-        ]:
-
-            response = self.get_response(
-                self.project.organization.slug,
-                field="sum(sentry.sessions.session)",
-                groupBy="environment",
-                query=query,
-            )
-
-            assert response.status_code == 400, query
+        assert response.status_code == 400, query
 
     @with_feature(FEATURE_FLAG)
     def test_valid_filter(self):
+        for tag in ("release", "environment"):
+            indexer.record(tag)
         query = "release:myapp@2.0.0"
         response = self.get_success_response(
             self.project.organization.slug,
@@ -331,27 +117,9 @@ class OrganizationMetricDataTest(APITestCase):
         """Order by tag is not supported (yet)"""
         response = self.get_response(
             self.project.organization.slug,
-            field="sum(sentry.sessions.session)",
+            field=["sum(sentry.sessions.session)", "environment"],
             groupBy="environment",
             orderBy="environment",
-        )
-        assert response.status_code == 400
-
-    @with_feature(FEATURE_FLAG)
-    def test_orderby_2(self):
-        """Support for more than one field is supported with order by"""
-        response = self.get_response(
-            self.project.organization.slug,
-            field=["sum(sentry.sessions.session)", "count_unique(sentry.sessions.user)"],
-            orderBy=["sum(sentry.sessions.session)"],
-        )
-        assert response.status_code == 200
-
-    @with_feature(FEATURE_FLAG)
-    def test_orderby_percentile(self):
-        """Order by tag is not supported yet"""
-        response = self.get_response(
-            self.project.organization.slug, field="p95(session)", orderBy="p95(session)"
         )
         assert response.status_code == 400
 
@@ -364,7 +132,6 @@ class OrganizationMetricDataTest(APITestCase):
         response = self.get_response(
             self.organization.slug,
             field="count(sentry.transactions.measurements.lcp)",
-            datasource="snuba",
             groupBy="transaction",
             per_page=2,
         )
@@ -382,12 +149,10 @@ class OrganizationMetricDataTest(APITestCase):
         response = self.get_response(
             self.organization.slug,
             field="count(sentry.transactions.measurements.lcp)",
-            datasource="snuba",
             groupBy="transaction",
             cursor=Cursor(0, 1),
         )
         assert response.status_code == 400
-        print(response.json())
         assert response.json()["detail"] == (
             "'cursor' is only supported in combination with 'orderBy'"
         )
@@ -401,15 +166,6 @@ class OrganizationMetricDataTest(APITestCase):
         )
         assert response.status_code == 400
 
-
-class OrganizationMetricIntegrationTest(SessionMetricsTestCase, APITestCase):
-    endpoint = "sentry-api-0-organization-metrics-data"
-
-    def setUp(self):
-        super().setUp()
-        self.project2 = self.create_project()
-        self.login_as(user=self.user)
-
     @with_feature(FEATURE_FLAG)
     def test_separate_projects(self):
         # Insert session metrics:
@@ -421,7 +177,6 @@ class OrganizationMetricIntegrationTest(SessionMetricsTestCase, APITestCase):
                 field="sum(sentry.sessions.session)",
                 statsPeriod="1h",
                 interval="1h",
-                datasource="snuba",
             )
             if project_id is not None:
                 kwargs["project"] = project_id
@@ -479,7 +234,6 @@ class OrganizationMetricIntegrationTest(SessionMetricsTestCase, APITestCase):
             query="measurement_rating:poor",
             statsPeriod="1h",
             interval="1h",
-            datasource="snuba",
             groupBy="transaction",
             orderBy="-count(sentry.transactions.measurements.lcp)",
             per_page=2,
@@ -531,7 +285,6 @@ class OrganizationMetricIntegrationTest(SessionMetricsTestCase, APITestCase):
             field="p50(sentry.transactions.measurements.lcp)",
             statsPeriod="1h",
             interval="1h",
-            datasource="snuba",
             groupBy="tag1",
             orderBy="p50(sentry.transactions.measurements.lcp)",
         )
@@ -581,7 +334,6 @@ class OrganizationMetricIntegrationTest(SessionMetricsTestCase, APITestCase):
             field="p50(sentry.transactions.measurements.lcp)",
             statsPeriod="1h",
             interval="1h",
-            datasource="snuba",
             groupBy="tag1",
             orderBy="p50(sentry.transactions.measurements.lcp)",
             per_page=1,
@@ -596,7 +348,6 @@ class OrganizationMetricIntegrationTest(SessionMetricsTestCase, APITestCase):
             field="p50(sentry.transactions.measurements.lcp)",
             statsPeriod="1h",
             interval="1h",
-            datasource="snuba",
             groupBy="tag1",
             orderBy="p50(sentry.transactions.measurements.lcp)",
             per_page=1,
@@ -642,7 +393,6 @@ class OrganizationMetricIntegrationTest(SessionMetricsTestCase, APITestCase):
             field="p50(sentry.transactions.measurements.lcp)",
             statsPeriod="1h",
             interval="1h",
-            datasource="snuba",
             groupBy="tag1",
             orderBy="p50(sentry.transactions.measurements.lcp)",
             per_page=1,
@@ -665,7 +415,6 @@ class OrganizationMetricIntegrationTest(SessionMetricsTestCase, APITestCase):
             ],
             statsPeriod="1h",
             interval="1h",
-            datasource="snuba",
             groupBy=["project_id", "transaction"],
             orderBy="p50(sentry.transactions.measurements.lcp)",
         )
@@ -690,7 +439,6 @@ class OrganizationMetricIntegrationTest(SessionMetricsTestCase, APITestCase):
             ],
             statsPeriod="1h",
             interval="1h",
-            datasource="snuba",
             groupBy=["project_id", "transaction"],
             orderBy="p50(sentry.transactions.measurements.lcp)",
         )
@@ -721,7 +469,6 @@ class OrganizationMetricIntegrationTest(SessionMetricsTestCase, APITestCase):
             ],
             statsPeriod="1h",
             interval="1h",
-            datasource="snuba",
             groupBy=["project_id", "transaction"],
             orderBy="p50(sentry.transactions.measurements.lcp)",
         )
@@ -787,7 +534,6 @@ class OrganizationMetricIntegrationTest(SessionMetricsTestCase, APITestCase):
             ],
             statsPeriod="1h",
             interval="1h",
-            datasource="snuba",
             groupBy=["project_id", "transaction"],
             orderBy="p50(sentry.transactions.measurements.lcp)",
         )
@@ -867,7 +613,6 @@ class OrganizationMetricIntegrationTest(SessionMetricsTestCase, APITestCase):
             ],
             statsPeriod="1h",
             interval="1h",
-            datasource="snuba",
             groupBy=["project_id", "transaction"],
             orderBy="p50(sentry.transactions.measurements.lcp)",
         )
@@ -1008,7 +753,6 @@ class OrganizationMetricIntegrationTest(SessionMetricsTestCase, APITestCase):
             ],
             statsPeriod="1h",
             interval="1h",
-            datasource="snuba",
             groupBy=["project_id", "transaction"],
             orderBy="p50(sentry.transactions.measurements.lcp)",
         )
@@ -1038,7 +782,6 @@ class OrganizationMetricIntegrationTest(SessionMetricsTestCase, APITestCase):
             interval="1h",
             field="sum(sentry.sessions.session)",
             groupBy=["project_id", "session.status"],
-            datasource="snuba",
         )
 
         assert response.status_code == 200
@@ -1071,7 +814,6 @@ class OrganizationMetricIntegrationTest(SessionMetricsTestCase, APITestCase):
             field="sum(sentry.sessions.session)",
             statsPeriod="1h",
             interval="1h",
-            datasource="snuba",
             groupBy=["session.status", "foo"],
         )
 
@@ -1084,7 +826,6 @@ class OrganizationMetricIntegrationTest(SessionMetricsTestCase, APITestCase):
             field="sum(sentry.sessions.session)",
             statsPeriod="1h",
             interval="1h",
-            datasource="snuba",
             groupBy=["session.status", "bar"],
         )
         assert response.status_code == 400
@@ -1100,7 +841,6 @@ class OrganizationMetricIntegrationTest(SessionMetricsTestCase, APITestCase):
             field="sum(sentry.sessions.session)",
             statsPeriod="1h",
             interval="1h",
-            datasource="snuba",
             query="foo:123",  # Unknown tag key
         )
         assert response.status_code == 400
@@ -1110,7 +850,6 @@ class OrganizationMetricIntegrationTest(SessionMetricsTestCase, APITestCase):
             field="sum(sentry.sessions.session)",
             statsPeriod="1h",
             interval="1h",
-            datasource="snuba",
             query="release:123",  # Unknown tag value is fine.
         )
 
@@ -1201,7 +940,6 @@ class OrganizationMetricsIndexIntegrationTest(OrganizationMetricMetaIntegrationT
 
         response = self.get_success_response(
             self.organization.slug,
-            datasource="snuba",  # TODO: remove datasource arg
         )
 
         assert response.data == [
@@ -1221,7 +959,6 @@ class OrganizationMetricDetailsIntegrationTest(OrganizationMetricMetaIntegration
         response = self.get_success_response(
             self.organization.slug,
             "metric1",
-            datasource="snuba",  # TODO: remove datasource arg
         )
         assert response.data == {
             "name": "metric1",
@@ -1239,7 +976,6 @@ class OrganizationMetricDetailsIntegrationTest(OrganizationMetricMetaIntegration
         response = self.get_success_response(
             self.organization.slug,
             "metric2",
-            datasource="snuba",  # TODO: remove datasource arg
         )
         assert response.data == {
             "name": "metric2",
@@ -1257,7 +993,6 @@ class OrganizationMetricDetailsIntegrationTest(OrganizationMetricMetaIntegration
         response = self.get_success_response(
             self.organization.slug,
             "metric3",
-            datasource="snuba",  # TODO: remove datasource arg
         )
         assert response.data == {
             "name": "metric3",
@@ -1276,7 +1011,6 @@ class OrganizationMetricsTagsIntegrationTest(OrganizationMetricMetaIntegrationTe
     def test_metric_tags(self):
         response = self.get_success_response(
             self.organization.slug,
-            datasource="snuba",  # TODO: remove datasource arg
         )
         assert response.data == [
             {"key": "tag1"},
@@ -1288,7 +1022,6 @@ class OrganizationMetricsTagsIntegrationTest(OrganizationMetricMetaIntegrationTe
         # When metric names are supplied, get intersection of tag names:
         response = self.get_success_response(
             self.organization.slug,
-            datasource="snuba",  # TODO: remove datasource arg
             metric=["metric1", "metric2"],
         )
         assert response.data == [
@@ -1298,7 +1031,6 @@ class OrganizationMetricsTagsIntegrationTest(OrganizationMetricMetaIntegrationTe
 
         response = self.get_success_response(
             self.organization.slug,
-            datasource="snuba",  # TODO: remove datasource arg
             metric=["metric1", "metric2", "metric3"],
         )
         assert response.data == []
@@ -1309,11 +1041,28 @@ class OrganizationMetricsTagDetailsIntegrationTest(OrganizationMetricMetaIntegra
     endpoint = "sentry-api-0-organization-metrics-tag-details"
 
     @with_feature(FEATURE_FLAG)
+    def test_unknown_tag(self):
+        indexer.record("bar")
+        response = self.get_success_response(self.project.organization.slug, "bar")
+        assert response.data == []
+
+    @with_feature(FEATURE_FLAG)
+    def test_non_existing_tag(self):
+        response = self.get_response(self.project.organization.slug, "bar")
+        assert response.status_code == 400
+
+    @with_feature(FEATURE_FLAG)
+    def test_non_existing_filter(self):
+        indexer.record("bar")
+        response = self.get_response(self.project.organization.slug, "bar", metric="bad")
+        assert response.status_code == 200
+        assert response.data == []
+
+    @with_feature(FEATURE_FLAG)
     def test_metric_tag_details(self):
         response = self.get_success_response(
             self.organization.slug,
             "tag1",
-            datasource="snuba",  # TODO: remove datasource arg
         )
         assert response.data == [
             {"key": "tag1", "value": "value1"},
@@ -1325,7 +1074,6 @@ class OrganizationMetricsTagDetailsIntegrationTest(OrganizationMetricMetaIntegra
             self.organization.slug,
             "tag1",
             metric=["metric1"],
-            datasource="snuba",  # TODO: remove datasource arg
         )
         assert response.data == [
             {"key": "tag1", "value": "value1"},
@@ -1336,6 +1084,5 @@ class OrganizationMetricsTagDetailsIntegrationTest(OrganizationMetricMetaIntegra
             self.organization.slug,
             "tag1",
             metric=["metric1", "metric2"],
-            datasource="snuba",  # TODO: remove datasource arg
         )
         assert response.data == []
